+++ Tests/Fixtures/config/nested_config_builder.php
@@ -1,9 +1,11 @@
 <?php
 
-use Symfony\Component\DependencyInjection\Tests\Fixtures\AcmeConfig;
-
 if ('prod' !== $env) {
     return;
 }
 
-return new AcmeConfig(['color' => 'red']);
+return [
+    'acme' => [
+        'color' => 'red',
+    ],
+];