+++ Tests/Fixtures/config/return_generator.php
@@ -1,7 +1,0 @@
-<?php
-
-use Symfony\Component\DependencyInjection\Tests\Fixtures\AcmeConfig;
-
-return function () {
-    yield new AcmeConfig(['color' => 'red']);
-};