+++ Tests/Fixtures/config/return_iterable_configs.php
@@ -1,8 +1,0 @@
-<?php
-
-use Symfony\Component\DependencyInjection\Tests\Fixtures\AcmeConfig;
-
-return [
-    'acme' => ['color' => 'red'],
-    new AcmeConfig(),
-];